@@ -1,4 +1,6 @@
-<?php declare(strict_types = 1);
+<?php
+
+declare(strict_types=1);
 
 namespace Quest\Macros;
 
@@ -8,10 +10,9 @@ class withMinimumRelevance
 {
     /**
      * Construct a fuzzy search expression.
-     *
      */
     public static function make(Builder $builder, int $score) : Builder
     {
-        return $builder->having('_fuzzy_relevance_', '>=',  $score);
+        return $builder->having('_fuzzy_relevance_', '>=', $score);
     }
 }
